+++ docs/changes.rst
@@ -39,6 +39,8 @@ To be released.
   <geofront.masterkey.PeriodicalRenewal>`, and :func:`regenerate()
   <geofront.regen.regenerate>`.
 - Added :class:`~geofront.backends.cloud.CloudKeyStore`.  [:issue:`2`]
+- Added :class:`~geofront.backends.cloud.CloudMasterPublicKeyStore`.
+  [:issue:`2`]
 
 .. _enum34: https://pypi.python.org/pypi/enum34
 .. _singledispatch: https://pypi.python.org/pypi/singledispatch